comment added
Hi Ramana,
See the section highlighted
I asked the lecturer about this, he advised not to load the whole file into memory as it won't work with the bigTwitter.json (the file is to big). Hence we have to read lead by line, see my latest commit. I was able to run it on spartan on the bigtwitter file and got the top 10 most frequent hashtags, however I am yet to develop it for languages.